+++ saleor/order/events.py
@@ -1,15 +1,14 @@
 from decimal import Decimal
 from typing import Dict, List, Optional, Tuple, Union
 
-from django.contrib.auth.base_user import AbstractBaseUser
-
-from ..account.models import Address
+from ..account import events as account_events
+from ..account.models import Address, User
 from ..order.models import Fulfillment, FulfillmentLine, Order, OrderLine
 from ..payment.models import Payment
 from . import OrderEvents, OrderEventsEmails
 from .models import OrderEvent
 
-UserType = AbstractBaseUser
+UserType = User
 
 
 def _lines_per_quantity_to_line_object_list(quantities_per_order_line):
@@ -29,17 +28,13 @@ def _get_payment_data(amount: Optional[Decimal], payment: Payment) -> Dict:
     }
 
 
-def _new_event(**data) -> OrderEvent:
-    return OrderEvent.objects.create(**data)
-
-
 def email_sent_event(
     *,
     order: Order,
     user: Optional[UserType],
     email_type: OrderEventsEmails,
     user_pk: int = None,
-):
+) -> OrderEvent:
 
     if user is not None and not user.is_anonymous:
         kwargs = {"user": user}
@@ -48,7 +43,7 @@ def email_sent_event(
     else:
         kwargs = {}
 
-    return _new_event(
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.EMAIL_SENT,
         parameters={"email": order.get_user_current_email(), "email_type": email_type},
@@ -56,19 +51,23 @@ def email_sent_event(
     )
 
 
-def email_resent_event(*, order: Order, user: UserType, email_type: OrderEventsEmails):
+def email_resent_event(
+    *, order: Order, user: UserType, email_type: OrderEventsEmails
+) -> OrderEvent:
     raise NotImplementedError
 
 
-def draft_order_created_event(*, order: Order, user: UserType):
-    return _new_event(order=order, type=OrderEvents.DRAFT_CREATED, user=user)
+def draft_order_created_event(*, order: Order, user: UserType) -> OrderEvent:
+    return OrderEvent.objects.create(
+        order=order, type=OrderEvents.DRAFT_CREATED, user=user
+    )
 
 
 def draft_order_added_products_event(
     *, order: Order, user: UserType, order_lines: List[Tuple[int, OrderLine]]
-):
+) -> OrderEvent:
 
-    return _new_event(
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.DRAFT_ADDED_PRODUCTS,
         user=user,
@@ -78,9 +77,9 @@ def draft_order_added_products_event(
 
 def draft_order_removed_products_event(
     *, order: Order, user: UserType, order_lines: List[Tuple[int, OrderLine]]
-):
+) -> OrderEvent:
 
-    return _new_event(
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.DRAFT_REMOVED_PRODUCTS,
         user=user,
@@ -88,19 +87,25 @@ def draft_order_removed_products_event(
     )
 
 
-def order_created_event(*, order: Order, user: UserType, from_draft=False):
-    event_type = OrderEvents.PLACED_FROM_DRAFT if from_draft else OrderEvents.PLACED
+def order_created_event(
+    *, order: Order, user: UserType, from_draft=False
+) -> OrderEvent:
+    if from_draft:
+        event_type = OrderEvents.PLACED_FROM_DRAFT
+    else:
+        event_type = OrderEvents.PLACED
+        account_events.customer_placed_order_event(user=user, order=order)
 
     if user.is_anonymous:
         user = None
 
-    return _new_event(order=order, type=event_type, user=user)
+    return OrderEvent.objects.create(order=order, type=event_type, user=user)
 
 
 def draft_order_oversold_items_event(
     *, order: Order, user: UserType, oversold_items: List[str]
-):
-    return _new_event(
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.OVERSOLD_ITEMS,
         user=user,
@@ -108,22 +113,24 @@ def draft_order_oversold_items_event(
     )
 
 
-def order_canceled_event(*, order: Order, user: UserType):
-    return _new_event(order=order, type=OrderEvents.CANCELED, user=user)
+def order_canceled_event(*, order: Order, user: UserType) -> OrderEvent:
+    return OrderEvent.objects.create(order=order, type=OrderEvents.CANCELED, user=user)
 
 
-def order_manually_marked_as_paid_event(*, order: Order, user: UserType):
-    return _new_event(order=order, type=OrderEvents.ORDER_MARKED_AS_PAID, user=user)
+def order_manually_marked_as_paid_event(*, order: Order, user: UserType) -> OrderEvent:
+    return OrderEvent.objects.create(
+        order=order, type=OrderEvents.ORDER_MARKED_AS_PAID, user=user
+    )
 
 
-def order_fully_paid_event(*, order: Order):
-    return _new_event(order=order, type=OrderEvents.ORDER_FULLY_PAID)
+def order_fully_paid_event(*, order: Order) -> OrderEvent:
+    return OrderEvent.objects.create(order=order, type=OrderEvents.ORDER_FULLY_PAID)
 
 
 def payment_captured_event(
     *, order: Order, user: UserType, amount: Decimal, payment: Payment
-):
-    return _new_event(
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.PAYMENT_CAPTURED,
         user=user,
@@ -133,8 +140,8 @@ def payment_captured_event(
 
 def payment_refunded_event(
     *, order: Order, user: UserType, amount: Decimal, payment: Payment
-):
-    return _new_event(
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.PAYMENT_REFUNDED,
         user=user,
@@ -142,8 +149,10 @@ def payment_refunded_event(
     )
 
 
-def payment_voided_event(*, order: Order, user: UserType, payment: Payment):
-    return _new_event(
+def payment_voided_event(
+    *, order: Order, user: UserType, payment: Payment
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.PAYMENT_VOIDED,
         user=user,
@@ -153,22 +162,22 @@ def payment_voided_event(*, order: Order, user: UserType, payment: Payment):
 
 def payment_failed_event(
     *, order: Order, user: UserType, message: str, payment: Payment
-):
+) -> OrderEvent:
 
     parameters = {"message": message}
 
     if payment:
         parameters.update({"gateway": payment.gateway, "payment_id": payment.token})
 
-    return _new_event(
+    return OrderEvent.objects.create(
         order=order, type=OrderEvents.PAYMENT_FAILED, user=user, parameters=parameters
     )
 
 
 def fulfillment_canceled_event(
     *, order: Order, user: UserType, fulfillment: Fulfillment
-):
-    return _new_event(
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.FULFILLMENT_CANCELED,
         user=user,
@@ -178,8 +187,8 @@ def fulfillment_canceled_event(
 
 def fulfillment_restocked_items_event(
     *, order: Order, user: UserType, fulfillment: Union[Order, Fulfillment]
-):
-    return _new_event(
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.FULFILLMENT_RESTOCKED_ITEMS,
         user=user,
@@ -189,8 +198,8 @@ def fulfillment_restocked_items_event(
 
 def fulfillment_fulfilled_items_event(
     *, order: Order, user: UserType, fulfillment_lines: List[FulfillmentLine]
-):
-    return _new_event(
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.FULFILLMENT_FULFILLED_ITEMS,
         user=user,
@@ -200,8 +209,8 @@ def fulfillment_fulfilled_items_event(
 
 def fulfillment_tracking_updated_event(
     *, order: Order, user: UserType, tracking_number: str, fulfillment: Fulfillment
-):
-    return _new_event(
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.TRACKING_UPDATED,
         user=user,
@@ -212,8 +221,13 @@ def fulfillment_tracking_updated_event(
     )
 
 
-def order_note_added_event(*, order: Order, user: UserType, message: str):
-    return _new_event(
+def order_note_added_event(*, order: Order, user: UserType, message: str) -> OrderEvent:
+    if order.user.pk == user.pk:
+        account_events.customer_added_to_note_order_event(
+            user=user, order=order, message=message
+        )
+
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.NOTE_ADDED,
         user=user,
@@ -221,8 +235,10 @@ def order_note_added_event(*, order: Order, user: UserType, message: str):
     )
 
 
-def order_updated_address_event(*, order: Order, user: UserType, address: Address):
-    return _new_event(
+def order_updated_address_event(
+    *, order: Order, user: UserType, address: Address
+) -> OrderEvent:
+    return OrderEvent.objects.create(
         order=order,
         type=OrderEvents.UPDATED_ADDRESS,
         user=user,